Moved expand_star to seperate file
