+++ go/vt/vtgate/planbuilder/route_planning_test.go
@@ -20,10 +20,6 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/stretchr/testify/require"
-
-	"vitess.io/vitess/go/sqltypes"
-
 	"vitess.io/vitess/go/vt/vtgate/semantics"
 
 	"github.com/stretchr/testify/assert"
@@ -125,155 +121,4 @@ func TestClone(t *testing.T) {
 	clonedRP.vindexPreds[0].foundVindex = &vindexes.Null{}
 	assert.NotNil(t, clonedRP.vindexPreds[0].foundVindex)
 	assert.Nil(t, original.vindexPreds[0].foundVindex)
-}
-
-func TestExpandStar(t *testing.T) {
-	schemaInfo := &fakeSI{
-		tables: map[string]*vindexes.Table{
-			"t1": {
-				Name: sqlparser.NewTableIdent("t1"),
-				Columns: []vindexes.Column{{
-					Name: sqlparser.NewColIdent("a"),
-					Type: sqltypes.VarChar,
-				}, {
-					Name: sqlparser.NewColIdent("b"),
-					Type: sqltypes.VarChar,
-				}, {
-					Name: sqlparser.NewColIdent("c"),
-					Type: sqltypes.VarChar,
-				}},
-				ColumnListAuthoritative: true,
-			},
-			"t2": {
-				Name: sqlparser.NewTableIdent("t2"),
-				Columns: []vindexes.Column{{
-					Name: sqlparser.NewColIdent("c1"),
-					Type: sqltypes.VarChar,
-				}, {
-					Name: sqlparser.NewColIdent("c2"),
-					Type: sqltypes.VarChar,
-				}},
-				ColumnListAuthoritative: true,
-			},
-			"t3": { // non authoritative table.
-				Name: sqlparser.NewTableIdent("t3"),
-				Columns: []vindexes.Column{{
-					Name: sqlparser.NewColIdent("col"),
-					Type: sqltypes.VarChar,
-				}},
-				ColumnListAuthoritative: false,
-			},
-		},
-	}
-	cDB := "db"
-	tcases := []struct {
-		sql    string
-		expSQL string
-		expErr string
-	}{{
-		sql:    "select * from t1",
-		expSQL: "select t1.a as a, t1.b as b, t1.c as c from t1",
-	}, {
-		sql:    "select t1.* from t1",
-		expSQL: "select t1.a as a, t1.b as b, t1.c as c from t1",
-	}, {
-		sql:    "select *, 42, t1.* from t1",
-		expSQL: "select t1.a as a, t1.b as b, t1.c as c, 42, t1.a as a, t1.b as b, t1.c as c from t1",
-	}, {
-		sql:    "select 42, t1.* from t1",
-		expSQL: "select 42, t1.a as a, t1.b as b, t1.c as c from t1",
-	}, {
-		sql:    "select * from t1, t2",
-		expSQL: "select t1.a as a, t1.b as b, t1.c as c, t2.c1 as c1, t2.c2 as c2 from t1, t2",
-	}, {
-		sql:    "select t1.* from t1, t2",
-		expSQL: "select t1.a as a, t1.b as b, t1.c as c from t1, t2",
-	}, {
-		sql:    "select *, t1.* from t1, t2",
-		expSQL: "select t1.a as a, t1.b as b, t1.c as c, t2.c1 as c1, t2.c2 as c2, t1.a as a, t1.b as b, t1.c as c from t1, t2",
-	}, { // aliased table
-		sql:    "select * from t1 a, t2 b",
-		expSQL: "select a.a as a, a.b as b, a.c as c, b.c1 as c1, b.c2 as c2 from t1 as a, t2 as b",
-	}, { // t3 is non-authoritative table
-		sql:    "select * from t3",
-		expSQL: "select * from t3",
-	}, { // t3 is non-authoritative table
-		sql:    "select * from t1, t2, t3",
-		expSQL: "select * from t1, t2, t3",
-	}, { // t3 is non-authoritative table
-		sql:    "select t1.*, t2.*, t3.* from t1, t2, t3",
-		expSQL: "select t1.a as a, t1.b as b, t1.c as c, t2.c1 as c1, t2.c2 as c2, t3.* from t1, t2, t3",
-	}, {
-		sql:    "select foo.* from t1, t2",
-		expErr: "Unknown table 'foo'",
-	}}
-	for _, tcase := range tcases {
-		t.Run(tcase.sql, func(t *testing.T) {
-			ast, err := sqlparser.Parse(tcase.sql)
-			require.NoError(t, err)
-			semTable, err := semantics.Analyze(ast, cDB, schemaInfo)
-			require.NoError(t, err)
-			expandedSelect, err := expandStar(ast.(*sqlparser.Select), semTable)
-			if tcase.expErr == "" {
-				require.NoError(t, err)
-				assert.Equal(t, tcase.expSQL, sqlparser.String(expandedSelect))
-			} else {
-				require.EqualError(t, err, tcase.expErr)
-			}
-		})
-	}
-}
-
-func TestSemTableDependenciesAfterExpandStar(t *testing.T) {
-	schemaInfo := &fakeSI{tables: map[string]*vindexes.Table{
-		"t1": {
-			Name: sqlparser.NewTableIdent("t1"),
-			Columns: []vindexes.Column{{
-				Name: sqlparser.NewColIdent("a"),
-				Type: sqltypes.VarChar,
-			}},
-			ColumnListAuthoritative: true,
-		}}}
-	tcases := []struct {
-		sql         string
-		expSQL      string
-		sameTbl     int
-		otherTbl    int
-		expandedCol int
-	}{{
-		sql:      "select a, * from t1",
-		expSQL:   "select a, t1.a as a from t1",
-		otherTbl: -1, sameTbl: 0, expandedCol: 1,
-	}, {
-		sql:      "select t2.a, t1.a, t1.* from t1, t2",
-		expSQL:   "select t2.a, t1.a, t1.a as a from t1, t2",
-		otherTbl: 0, sameTbl: 1, expandedCol: 2,
-	}, {
-		sql:      "select t2.a, t.a, t.* from t1 t, t2",
-		expSQL:   "select t2.a, t.a, t.a as a from t1 as t, t2",
-		otherTbl: 0, sameTbl: 1, expandedCol: 2,
-	}}
-	for _, tcase := range tcases {
-		t.Run(tcase.sql, func(t *testing.T) {
-			ast, err := sqlparser.Parse(tcase.sql)
-			require.NoError(t, err)
-			semTable, err := semantics.Analyze(ast, "", schemaInfo)
-			require.NoError(t, err)
-			expandedSelect, err := expandStar(ast.(*sqlparser.Select), semTable)
-			require.NoError(t, err)
-			assert.Equal(t, tcase.expSQL, sqlparser.String(expandedSelect))
-			if tcase.otherTbl != -1 {
-				assert.NotEqual(t,
-					semTable.Dependencies(expandedSelect.SelectExprs[tcase.otherTbl].(*sqlparser.AliasedExpr).Expr),
-					semTable.Dependencies(expandedSelect.SelectExprs[tcase.expandedCol].(*sqlparser.AliasedExpr).Expr),
-				)
-			}
-			if tcase.sameTbl != -1 {
-				assert.Equal(t,
-					semTable.Dependencies(expandedSelect.SelectExprs[tcase.sameTbl].(*sqlparser.AliasedExpr).Expr),
-					semTable.Dependencies(expandedSelect.SelectExprs[tcase.expandedCol].(*sqlparser.AliasedExpr).Expr),
-				)
-			}
-		})
-	}
 }
